chore(license): update to GNU Affero General Public License (AGPL)
The project license has been updated to the GNU Affero General Public License (AGPL) to ensure users and developers can freely use, modify, and distribute the code under the terms of this more restrictive license. This also ensures that any modifications or use of the software in a networked environment are made available to the public.